Consolidate pay/withdraw request data structs
