+++ libs/sdk-core/src/bridge_generated.rs
@@ -51,7 +51,6 @@ use crate::models::GreenlightNodeConfig;
 use crate::models::HealthCheckStatus;
 use crate::models::ListPaymentsRequest;
 use crate::models::LnPaymentDetails;
-use crate::models::LnUrlPayRequest;
 use crate::models::LnUrlWithdrawRequest;
 use crate::models::LogEntry;
 use crate::models::MaxReverseSwapAmountResponse;
@@ -112,6 +111,7 @@ use crate::LnUrlAuthRequestData;
 use crate::LnUrlCallbackStatus;
 use crate::LnUrlErrorData;
 use crate::LnUrlPayErrorData;
+use crate::LnUrlPayRequest;
 use crate::LnUrlPayRequestData;
 use crate::LnUrlPayResult;
 use crate::LnUrlPaySuccessData;
